Some tweaks to help with deployment
