+++ claudia.json
@@ -6,6 +6,6 @@
   },
   "api": {
     "id": "raxyggo1e4",
-    "module": "index"
+    "module": "lib/bot"
   }
 }
